Move methods and properties to EA Mixin classes (#21776)
