+++ pandas/core/arrays/timedelta.py
@@ -2,12 +2,27 @@
 
 from pandas._libs import tslib
 from pandas._libs.tslib import Timedelta, NaT
+from pandas._libs.tslibs.fields import get_timedelta_field
 
 from pandas.core.dtypes.common import _TD_DTYPE
 
 from pandas.tseries.offsets import Tick
 
 from .datetimelike import DatetimeLikeArrayMixin
+
+
+def _field_accessor(name, alias, docstring=None):
+    def f(self):
+        values = self.asi8
+        result = get_timedelta_field(values, alias)
+        if self.hasnans:
+            result = self._maybe_mask_results(result, convert='float64')
+
+        return result
+
+    f.__name__ = name
+    f.__doc__ = docstring
+    return property(f)
 
 
 class TimedeltaArrayMixin(DatetimeLikeArrayMixin):
@@ -46,3 +61,15 @@ class TimedeltaArrayMixin(DatetimeLikeArrayMixin):
         datetimes : ndarray
         """
         return tslib.ints_to_pytimedelta(self.asi8)
+
+    days = _field_accessor("days", "days",
+                           " Number of days for each element. ")
+    seconds = _field_accessor("seconds", "seconds",
+                              " Number of seconds (>= 0 and less than 1 day) "
+                              "for each element. ")
+    microseconds = _field_accessor("microseconds", "microseconds",
+                                   "\nNumber of microseconds (>= 0 and less "
+                                   "than 1 second) for each\nelement. ")
+    nanoseconds = _field_accessor("nanoseconds", "nanoseconds",
+                                  "\nNumber of nanoseconds (>= 0 and less "
+                                  "than 1 microsecond) for each\nelement.\n")
